// Front-campain-app-main/campain-app-develop/src/types/campaign.types.ts
export interface CibleAPIResponse {
  id?: string;
  id_cible?: string;
  nom_cible: string;
  source: 'DB' | 'FILE';
  created_at?: string;
}
// ...
export interface Cible {
  id: string;
  code: string;
  name: string;
  type: string;
  timestamp: string;
}